Revert "feat: 회원 정보 조회 API 추가"
