--- src/main/java/co/kr/allpick/domain/member/service/MemberService.java
@@ -1,7 +1,0 @@
-package co.kr.allpick.domain.member.service;
-
-import co.kr.allpick.domain.member.dto.MemberResponseDto;
-
-public interface MemberService {
-    MemberResponseDto getMember(Long memberId);
-}
