--- src/main/java/co/kr/allpick/domain/member/service/impl/MemberServiceImpl.java
@@ -1,30 +1,0 @@
-package co.kr.allpick.domain.member.service.impl;
-
-import co.kr.allpick.domain.member.dto.MemberResponseDto;
-import co.kr.allpick.domain.member.repository.MemberRepository;
-import co.kr.allpick.domain.member.service.MemberService;
-import co.kr.allpick.global.exception.BusinessException;
-import co.kr.allpick.global.exception.ErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class MemberServiceImpl implements MemberService {
-
-    private static final Logger logger = LogManager.getLogger(MemberServiceImpl.class);
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public MemberResponseDto getMember(Long memberId) {
-        logger.info("회원 정보 조회 - memberId: {}", memberId);
-        return memberRepository.findById(memberId)
-                .map(MemberResponseDto::from)
-                .orElseThrow(() -> new BusinessException(ErrorCode.MEMBER_NOT_FOUND));
-    }
-}
